refactor: 토큰 확인 로직 수정
