+++ src/main/java/gift/auth/LoginMemberArgumentResolver.java
@@ -1,8 +1,5 @@
-package gift.common.validation;
+package gift.auth;
 
-import gift.member.domain.Member;
-import gift.member.exception.MemberAuthorizedErrorException;
-import gift.member.persistence.MemberRepository;
 import jakarta.servlet.http.HttpServletRequest;
 import org.springframework.core.MethodParameter;
 import org.springframework.web.bind.support.WebDataBinderFactory;
@@ -11,12 +8,6 @@ import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.method.support.ModelAndViewContainer;
 
 public class LoginMemberArgumentResolver implements HandlerMethodArgumentResolver {
-    private final MemberRepository memberRepository;
-
-    public LoginMemberArgumentResolver(MemberRepository memberRepository) {
-        this.memberRepository = memberRepository;
-    }
-
     @Override
     public boolean supportsParameter(MethodParameter parameter) {
         return parameter.getParameterAnnotation(LoginMember.class) != null;
@@ -30,9 +21,6 @@ public class LoginMemberArgumentResolver implements HandlerMethodArgumentResolve
             throw new IllegalArgumentException("Request 정보가 존재하지 않습니다.");
         }
 
-        Member member = memberRepository.findByUsername((String) request.getAttribute("username"))
-                .orElseThrow(MemberAuthorizedErrorException::new);
-
-        return member;
+        return request.getAttribute("id");
     }
 }
