+++ src/main/java/gift/config/WebConfig.java
@@ -3,7 +3,6 @@ package gift.config;
 import gift.auth.AuthenticationInterceptor;
 import gift.auth.AuthorizationInterceptor;
 import gift.auth.LoginMemberArgumentResolver;
-import gift.member.persistence.MemberRepository;
 import gift.member.service.JwtProvider;
 import java.util.List;
 import org.springframework.context.annotation.Configuration;
@@ -14,11 +13,9 @@ import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 @Configuration
 public class WebConfig implements WebMvcConfigurer {
     private final JwtProvider jwtProvider;
-    private final MemberRepository memberRepository;
 
-    public WebConfig(JwtProvider jwtProvider, MemberRepository memberRepository) {
+    public WebConfig(JwtProvider jwtProvider) {
         this.jwtProvider = jwtProvider;
-        this.memberRepository = memberRepository;
     }
 
     @Override
@@ -29,6 +26,6 @@ public class WebConfig implements WebMvcConfigurer {
 
     @Override
     public void addArgumentResolvers(List<HandlerMethodArgumentResolver> resolvers) {
-        resolvers.add(new LoginMemberArgumentResolver(memberRepository));
+        resolvers.add(new LoginMemberArgumentResolver());
     }
 }
